requirements.txt
asgiref==3.6.0
black==23.3.0
click==8.1.3
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
